refactor(web): address review comments on PR #1025
- Extract findTaskInSnapshots to task-select-helpers.ts (shared between
  desktop sidebar and mobile sheet). Mobile passes kanban.tasks as the
  optional fallback so it still locates tasks that arrived via WS before
  the snapshot fetch landed.
- Build workflowNameById from the already-scoped `workflows` array instead
  of the raw, unscoped kanbanMulti.snapshots. Stops the memo from re-running
  on snapshot updates from other workspaces and removes the risk of a stale
  workflow name leaking across a workspace switch. Same fix in both the
  desktop sidebar and the mobile sheet.
- Collapse the duplicated workflow filter in useWorkspaceSidebarTasks into
  a single `filteredWorkflows` memo.
